Redesign classes to focus on TrustAnchorRepository
