+++ webauthn4j-metadata/src/main/java/com/webauthn4j/metadata/MetadataStatementRepository.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2002-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.webauthn4j.metadata;
+
+import com.webauthn4j.data.attestation.authenticator.AAGUID;
+import com.webauthn4j.metadata.data.statement.MetadataStatement;
+
+import java.util.Set;
+
+public interface MetadataStatementRepository {
+
+    Set<MetadataStatement> find(AAGUID aaguid);
+    Set<MetadataStatement> find(byte[] attestationCertificateKeyIdentifier);
+}
